+++ builders/StrBuilder.js
@@ -8,10 +8,22 @@ export default class StrBuilder {
         this.value = value;
         return this;
     }
+    toLowerCase(combine) {
+        this.value = this.value.toLowerCase();
+        if (isNotEmpty(combine))
+            return this;
+        return this.value;
+    }
     toPascalCase(combine) {
         this.value = defineValue(this.value.match(/[A-Z]{2,}(?=[A-Z][a-z]+[0-9]*|\b)|[A-Z]?[a-z]+[0-9]*|[A-Z]|[0-9]+/g), [])
             .map((x) => x.slice(0, 1).toUpperCase() + x.slice(1).toLowerCase())
             .join("");
+        if (isNotEmpty(combine))
+            return this;
+        return this.value;
+    }
+    toUpperCase(combine) {
+        this.value = this.value.toUpperCase();
         if (isNotEmpty(combine))
             return this;
         return this.value;