+++ src/builders/StrBuilder.ts
@@ -13,10 +13,26 @@ export default class StrBuilder {
         return this;
     }
 
+    public toLowerCase(combine?: boolean): StrBuilder | string {
+        this.value = this.value.toLowerCase();
+
+        if (isNotEmpty(combine)) return this;
+
+        return this.value;
+    }
+
     public toPascalCase(combine?: boolean): StrBuilder | string {
         this.value = defineValue(this.value.match(/[A-Z]{2,}(?=[A-Z][a-z]+[0-9]*|\b)|[A-Z]?[a-z]+[0-9]*|[A-Z]|[0-9]+/g), [])
             .map((x: string) => x.slice(0, 1).toUpperCase() + x.slice(1).toLowerCase())
             .join("");
+
+        if (isNotEmpty(combine)) return this;
+
+        return this.value;
+    }
+
+    public toUpperCase(combine?: boolean): StrBuilder | string {
+        this.value = this.value.toUpperCase();
 
         if (isNotEmpty(combine)) return this;