+++ CHANGELOG.md
@@ -3,6 +3,22 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [v0.1.13](https://github.com/crenata/bejibun-utils/compare/v0.1.0...v0.1.13) - 2025-10-19
+
+### 🩹 Fixes
+
+### 📖 Changes
+What's New :
+- Adding `Str.toLowerCase()` & `Str.toUpperCase()`
+
+### ❤️Contributors
+- Havea Crenata ([@crenata](https://github.com/crenata))
+- Ghulje ([@ghulje](https://github.com/ghulje))
+
+**Full Changelog**: https://github.com/crenata/bejibun-utils/blob/master/CHANGELOG.md
+
+---
+
 ## [v0.1.0](https://github.com/crenata/bejibun-utils/compare/v0.1.0...v0.1.0) - 2025-10-16
 
 ### 🩹 Fixes